requirements.txt
fastapi
uvicorn
sqlalchemy
langchain
langgraph
python-dotenv
redis
pinecone
pinecone-text
qdrant-client
weaviate-client
pymilvus
pypdf
langchain_pinecone
langchain_community
google.genai
python-multipart==0.0.20
pydantic[email]
redis-db
scikit-learn